feat: 학생 관계 점수 계산
1. Scheduler 설정
2. 03:00에 학생 관계 점수 계산하는 로직 작성
3. RelationShip Entity relationPoint Double 로 변경

+++ backEnd/src/main/java/com/quiz/ourclass/domain/organization/service/MemberOrgServiceImpl.java
@@ -84,7 +84,7 @@ public class MemberOrgServiceImpl implements MemberOrgService {
         return RelationResponse.builder()
             .memberId(targetMember.getId())
             .memberName(targetMember.getName())
-            .relationPoint(relationship.getRelationPoint())
+            .relationPoint((int) relationship.getRelationPoint())
             .groupCount(relationship.getDesignGroupCount() + relationship.getFreeGroupCount())
             .tagGreetingCount(relationship.getTagGreetingCount())
             .receiveCount(receiveCount)
@@ -156,7 +156,7 @@ public class MemberOrgServiceImpl implements MemberOrgService {
                 MemberSimpleDTO memberSimpleDTO = memberMapper.memberToMemberSimpleDTO(target);
                 return RelationSimpleResponse.builder()
                     .member(memberSimpleDTO)
-                    .relationPoint(relationship.getRelationPoint())
+                    .relationPoint((int) relationship.getRelationPoint())
                     .build();
             }).toList();
     }
